pyproject.toml
[tool.poetry]
name = "attack-range"
version = "3.1.2"
description = ""
authors = ["P4T12ICK <pbareiss@splunk.com>"]

[tool.poetry.dependencies]
python = "^3.10"
ansible = "<10.0.0"
pywinrm = "^0.5.0"
PyYAML = "^6.0.2"
colorama = "^0.4.6"
python-terraform = "^0.10.1"
tabulate = "^0.9.0"
boto3 = "^1.35.27"
ansible-runner = "^2.4.0"
azure-mgmt-resource = "^23.1.1"
azure-mgmt-network = "^27.0.0"
azure-mgmt-compute = "^33.0.0"
azure-identity = "^1.18.0"
splunk-sdk = "^2.0.2"
urllib3 = "^2.2.3"
pycryptodome = "^3.20.0"
questionary = "^2.0.1"
Jinja2 = "^3.1.4"
python-vagrant = "^1.0.0"
gitpython = "^3.1.43"
google-cloud-compute = "*"
google-cloud-storage = "*"
google-cloud-bigtable = "*"
google-cloud-secret-manager = "*"
google-auth = "*"

[[tool.poetry.source]]
name = "pypi-public"
url = "https://pypi.org/simple/"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
